[FLINK-8302][table]Minor update to some typo fix and test cases added

+++ flink-libraries/flink-table/src/main/scala/org/apache/flink/table/functions/sql/ScalarSqlFunctions.scala
@@ -63,7 +63,7 @@ object ScalarSqlFunctions {
     SqlKind.OTHER_FUNCTION,
     ReturnTypes.ARG0,
     null,
-    OperandTypes.family(SqlTypeFamily.NUMERIC, SqlTypeFamily.INTEGER),
+    OperandTypes.family(SqlTypeFamily.INTEGER, SqlTypeFamily.INTEGER),
     SqlFunctionCategory.NUMERIC
   )
 
@@ -72,7 +72,7 @@ object ScalarSqlFunctions {
     SqlKind.OTHER_FUNCTION,
     ReturnTypes.ARG0,
     null,
-    OperandTypes.family(SqlTypeFamily.NUMERIC, SqlTypeFamily.INTEGER),
+    OperandTypes.family(SqlTypeFamily.INTEGER, SqlTypeFamily.INTEGER),
     SqlFunctionCategory.NUMERIC
   )
